+++ Fmt/rte_main_fmt.h
@@ -1,5 +1,5 @@
-#ifndef RTE_RTE_MAIN_H
-#define RTE_RTE_MAIN_H
+#ifndef RTE_RTE_MAIN_FMT_H
+#define RTE_RTE_MAIN_FMT_H
 /* rte_main_fmt.h - the main format definition file */
 
 /**** System message definitions - mandatory include (must be the first one) ****/