+++ Portable/CPU/Generic/rtedbg_generic_irq_disable.h
@@ -5,34 +5,37 @@
  */
 
 /*******************************************************************************
- * @file    rtedbg_cortex_m.h
+ * @file    rtedbg_generic_irq_disable.h
  * @author  Branko Premzel
  * @version RTEdbg library <DEVELOPMENT BRANCH>
  *
+ * @brief  Circular buffer space reservation using interrupt disable / enable.
+ *         => Check for the processor core-specific version, if available. 
+ *         It may be more optimized and faster than this generic one.
+ *         See also description in the Readme.md file.
  *
- * @brief  ARM Cortex core specific functions for the buffer space reservation.
- *         This version is suitable for ARM Cortex cores that do not support
- *         mutex instructions (e.g. M0/M0+/M23). Use 'rtedbg_cortex_m_mutex.h'
- *         for all cores that support mutex instructions. Do not use this file for
- *         Cortex M cores that support privileged mode since the interrupt disable
- *         has no effect if used in unprivileged software, e.g. in a RTOS task.
+ *         Do not use this file for cores that support privileged mode since the
+ *         interrupt disable has no effect if used in unprivileged software (such
+ *         as the ARM Cortex CPU cores), e.g. in a RTOS task.
+ *
  *         Interrupts are disabled for only a few CPU cycles while space is
  *         reserved in circular memory.
+ *
+ * The following macros have to be added to the rtedbg_config.h project file:
+ *  #define RTE_ENTER_CRITICAL() // Save status indicating whether interrupts are enabled and disable them
+ *  #define RTE_EXIT_CRITICAL()  // Enable interrupts again, if they have been enabled before.
+ * 
+ * Example for the ARM Cortex M0/M0+/M23:
+ *  #define RTE_ENTER_CRITICAL()  uint32_t irq_tmp = __get_PRIMASK(); __disable_irq();
+ *  #define RTE_EXIT_CRITICAL()   if (irq_tmp == 0U) { __enable_irq(); }
  *
  * @note   RTE_RESERVE_SPACE is defined as a macro instead of an inline function
  *         because the compiler typically generates smaller code when single-shot
  *         logging is enabled.
  ******************************************************************************/
 
-/********************************************************************************
- * @brief  Reserve space in the circular buffer. Since the Cortex M0 and M0+ cores do
- *         not support mutex instructions, interrupts should only be disabled for the
- *         duration of the space reservation and not while copying data to the buffer.
- *         Interrupts are therefore only disabled for a very short time.
- ********************************************************************************/
-
-#ifndef RTEDBG_CORTEX_M_H
-#define RTEDBG_CORTEX_M_H
+#ifndef RTEDBG_GENERIC_IRQ_DISABLE_H
+#define RTEDBG_GENERIC_IRQ_DISABLE_H
 
 #if RTE_SINGLE_SHOT_ENABLED == 0
 
@@ -41,18 +44,14 @@
  */
 #define RTE_RESERVE_SPACE(ptr, buf_idx, size)                        \
 do {                                                                 \
-    uint32_t primask = __get_PRIMASK();                              \
-    __disable_irq();                                                 \
+    RTE_ENTER_CRITICAL()                                             \
     buf_idx = ptr->buf_index;                                        \
     RTE_LIMIT_INDEX(buf_idx)                                         \
     ptr->buf_index = buf_idx + (size);                               \
-    if (primask == 0U)                                               \
-    {                                                                \
-        __enable_irq();                                              \
-    }                                                                \
+    RTE_EXIT_CRITICAL()                                              \
 } while(0)
 
-#else   /* RTE_SINGLE_SHOT_ENABLED == 0 */
+#else   /* RTE_SINGLE_SHOT_ENABLED == 1 */
 
 /* Single shot and post mortem / streaming data logging are possible.
  * Post mortem logging is the default mode. Single shot logging must be
@@ -60,8 +59,7 @@ do {                                                                 \
  */
 #define RTE_RESERVE_SPACE(ptr, buf_idx, size)                        \
 do {                                                                 \
-    uint32_t primask = __get_PRIMASK();                              \
-    __disable_irq();                                                 \
+    RTE_ENTER_CRITICAL()                                             \
     buf_idx = ptr->buf_index;                                        \
     if (ptr->rte_cfg & RTE_SINGLE_SHOT_LOGGING_IS_ACTIVE)            \
     {                                                                \
@@ -69,23 +67,17 @@ do {                                                                 \
         if ((buf_idx + (size)) >= (uint32_t)(RTE_BUFFER_SIZE))       \
         {                                                            \
             RTE_STOP_MESSAGE_LOGGING();                              \
-            if (primask == 0U)                                       \
-            {                                                        \
-               __enable_irq();                                       \
-            }                                                        \
+            RTE_EXIT_CRITICAL()                                      \
             return;        /* and exit the __rte_msg function. */    \
         }                                                            \
     }                                                                \
     RTE_LIMIT_INDEX(buf_idx)                                         \
     ptr->buf_index = buf_idx + (size);                               \
-    if (primask == 0U)                                               \
-    {                                                                \
-        __enable_irq();                                              \
-    }                                                                \
+    RTE_EXIT_CRITICAL()                                              \
 } while(0)
 #endif /* RTE_SINGLE_SHOT_ENABLED == 0 */
 
-#endif  // RTEDBG_CORTEX_M_H
+#endif  // RTEDBG_GENERIC_IRQ_DISABLE_H
 
 /*==== End of file ====*/