+++ Portable/Timer/STM32/rtedbg_timer_stm32L4_tim2.h
@@ -35,8 +35,8 @@ extern "C" {
 #if RTE_USE_LONG_TIMESTAMP != 0
 struct _tstamp64
 {
-    uint32_t l;    // Bottom and top part of the 64-bit timestamp
-    uint32_t h;
+    uint32_t l;    // Bottom part of the 64-bit timestamp
+    uint32_t h;    // Top part of the 64-bit timestamp
 } t_stamp;
 #endif // RTE_USE_LONG_TIMESTAMP != 0
 
@@ -80,7 +80,7 @@ __STATIC_FORCEINLINE uint32_t rte_get_timestamp(void)
 #if RTE_USE_LONG_TIMESTAMP != 0
 
 /*********************************************************************************
- * @brief  Writes a message with long time stamp to the buffer.
+ * @brief  Writes a message with long timestamp to the buffer.
  *         The low bits of the timestamp are included in the message words with the
  *         format ID. Only the higher 32 bits are transmitted in the message's
  *         data part.